Added BaseActivity.java class - Added attachBaseContext to BaseActivity - Added onOptionsItemSelected to BaseActivity - Removed attachBaseContext in many activities - Removed onOptionsItemSelected in many activities - Updated extends AppCompatActivity into BaseActivity in many activites - Added extends AppCompatActivity to BaseActivity
