+++ app/src/main/java/com/sakarsh/akarshseggemuresume/BaseActivity.java
@@ -1,0 +1,24 @@
+package com.sakarsh.akarshseggemuresume;
+
+import android.content.Context;
+import android.support.v7.app.AppCompatActivity;
+import android.view.MenuItem;
+
+public abstract class BaseActivity extends AppCompatActivity {
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+//            handles the click for back button
+            case android.R.id.home:
+                onBackPressed();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
+    @Override
+    protected void attachBaseContext(Context base) {
+        super.attachBaseContext(LocaleHelper.onAttach(base));
+    }
+}
